#57 Fixed binary blob handling with MariaDB Connector/J 3.5.2+
Root cause is that the driver returns java.sql.Blob as the column class
name in ResultSetMetaData while returning the contents directly as byte
arrays when fetched.

Since this behavior is so database and driver specific and causes issues
in instantiator lookup, minimally interfering solution is to use dialect
to provide type overrides to ResultSetUtils to make sure that the
expected and actually received types match.